backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "zaid",email = "zaid.al-bardan@tomtom.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4"
dependencies = [
    "flask (>=3.1.0,<3.2.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "google-api-python-client (>=2.164.0,<3.0.0)",
    "google-auth (>=2.38.0,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "email-validator (>=2.0.0,<3.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pydantic (==2.11.3)",
    "flask-jwt-extended (>=4.7.1,<5.0.0)",
    "geoalchemy2 (>=0.17.1,<0.18.0)",
    "shapely (>=2.1.1,<3.0.0)",
    "scikit-learn (>=1.7.0,<2.0.0)",
    "pyproj (>=3.7.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
